Improve handling of config file templates and include mime.types.
Move the base templates in to /etc/deploy/templates/base/.  This makes
it easier to extend them.  When given the name of a template to load
(e.g. nginx.conf.jinja), the jinja environment will first look in
/app/docker/templates/ and then in /etc/deploy/templates/.  It will
search for the name of the given template and then
'base/<templatename>'.  This allows templates to be overridden by
putting, say, nginx.conf.jinja in /app/docker/templates/ and also extend
the base templates by using "{% extends 'base/<templatename>' %}".

Allow additional environment variables to be used within the
user-defined templates.  All environment variables starting with 'USER_'
will be included in the variables access to any of the templates.

Create jinja blocks in the nginx base template to make it easier to
extend.

Include /etc/mime.types in the nginx config so that static files get sent
with the correct MIME type.