Fix security issue in gunicorn configuration.
By default, gunicorn allows X-Forwarded-Protocol, X-Forwarded-Proto, and
X-Forwarded-Ssl headers to be used to determine if the request is over a
secure protocol and therefore set wsgi.url_scheme accordingly. Our nginx
config only sets the X-Forwarded-Proto header without removing the other
2. This could allow an attacker to manually send one of the others to
make the application think that the request was sent over a secure
protocol when it actually wasn't.