Fix nginx config to set the Host header properly.
Since we're running this in a docker container where the port that nginx
is listening on might be different that what the user is using to
connect to it, we want to use the "Host" header that they sent when
proxying the request to gunicorn so that if the application needs to
compare the HTTP_REFERER to the HOST, they will have the same
server/port. We can't use $server_port because it's not going to be
what the client sends in the HTTP_REFERER header if the port is exposed
as a different port on the docker host.